# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "BicCD api文档",
        "contact": {
            "name": "ZiJiaGao",
            "email": "boringmanman@qq.com"
        },
        "version": "0.1"
    },
    "host": "bic-cd.farmergao.cn",
    "paths": {
        "/bic-cd/admin/v1/admin/": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth-Service"
                ],
                "summary": "创建管理员",
                "parameters": [
                    {
                        "description": "body 参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.RequestPostAdmin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gen.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.ResponsePostAdmin"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "header need Authorization data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "no api permission or no obj permission",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/bic-cd/admin/v1/auth/": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth-Service"
                ],
                "summary": "获取token",
                "parameters": [
                    {
                        "description": "body 参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.RequestAuth"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gen.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.ResponseAuth"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "header need Authorization data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "no api permission or no obj permission",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/bic-cd/manager/v1/service/:id/deploy": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manager-Service"
                ],
                "summary": "部署服务",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "some id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body 参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.RequestPostService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gen.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.ResponsePostService"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "header need Authorization data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "no api permission or no obj permission",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/bic-cd/manager/v1/services/": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manager-Service"
                ],
                "summary": "获取服务列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "参数无注释",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "参数无注释",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gen.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.ResponseGetService"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "header need Authorization data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "no api permission or no obj permission",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manager-Service"
                ],
                "summary": "添加服务",
                "parameters": [
                    {
                        "description": "body 参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.RequestPostService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gen.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.ResponsePostService"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "header need Authorization data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "no api permission or no obj permission",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.RequestAuth": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "description": "@gotags: validate:\"required\"",
                    "type": "string"
                },
                "username": {
                    "description": "@gotags: validate:\"required\"",
                    "type": "string"
                }
            }
        },
        "api.RequestPostAdmin": {
            "type": "object",
            "properties": {
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "username": {
                    "description": "账号",
                    "type": "string"
                }
            }
        },
        "api.RequestPostService": {
            "type": "object",
            "properties": {
                "config": {
                    "description": "配置文件",
                    "type": "string"
                },
                "description": {
                    "description": "服务描述",
                    "type": "string"
                },
                "exec_start": {
                    "description": "启动命令",
                    "type": "string"
                },
                "name": {
                    "description": "服务名",
                    "type": "string"
                },
                "port_max": {
                    "description": "最大端口号",
                    "type": "integer"
                },
                "port_min": {
                    "description": "最小端口号",
                    "type": "integer"
                },
                "user": {
                    "description": "运行用户",
                    "type": "string"
                },
                "working_dir": {
                    "description": "工作目录",
                    "type": "string"
                }
            }
        },
        "api.ResponseAuth": {
            "type": "object",
            "properties": {
                "token": {
                    "description": "token",
                    "type": "string"
                },
                "token_refresh": {
                    "description": "刷新token",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "api.ResponseGetService": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ServiceItem"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "api.ResponsePostAdmin": {
            "type": "object"
        },
        "api.ResponsePostService": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "api.ServiceItem": {
            "type": "object",
            "properties": {
                "config": {
                    "description": "配置文件",
                    "type": "string"
                },
                "description": {
                    "description": "服务描述",
                    "type": "string"
                },
                "exec_start": {
                    "description": "启动命令",
                    "type": "string"
                },
                "id": {
                    "description": "id",
                    "type": "integer"
                },
                "name": {
                    "description": "服务名",
                    "type": "string"
                },
                "port_max": {
                    "description": "最大端口号",
                    "type": "integer"
                },
                "port_min": {
                    "description": "最小端口号",
                    "type": "integer"
                },
                "user": {
                    "description": "运行用户",
                    "type": "string"
                },
                "working_dir": {
                    "description": "工作目录",
                    "type": "string"
                }
            }
        },
        "gen.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "detail": {
                    "type": "string"
                }
            }
        }
    }
}